change VSS type int to uint64 to fix when VSS > MAX(int)

+++ proc_stat.go
@@ -82,9 +82,9 @@ type ProcStat struct {
 	// in clock ticks.
 	Starttime uint64
 	// Virtual memory size in bytes.
-	VSize int
+	VSize uint64
 	// Resident set size in pages.
-	RSS int
+	RSS uint64
 
 	fs FS
 }
@@ -151,13 +151,13 @@ func (p Proc) NewStat() (ProcStat, error) {
 }
 
 // VirtualMemory returns the virtual memory size in bytes.
-func (s ProcStat) VirtualMemory() int {
+func (s ProcStat) VirtualMemory() uint64 {
 	return s.VSize
 }
 
 // ResidentMemory returns the resident memory size in bytes.
-func (s ProcStat) ResidentMemory() int {
-	return s.RSS * os.Getpagesize()
+func (s ProcStat) ResidentMemory() uint64 {
+	return s.RSS * uint64(os.Getpagesize())
 }
 
 // StartTime returns the unix timestamp of the process in seconds.
